# server/docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Daily Sumire Sync API",
    "description": "REST API specification for Daily Sumire - Offline-First Planner & Health Assistant with Dual-Token JWT Auth and Two-Way Synchronization.",
    "version": "1.8.58",
    "contact": {
      "name": "Daily Sumire Team"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3001/api",
      "description": "Local Development Server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "User": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "email": { "type": "string", "format": "email" },
          "firstName": { "type": "string" },
          "lastName": { "type": "string", "nullable": true },
          "username": { "type": "string", "nullable": true },
          "avatarUrl": { "type": "string", "nullable": true },
          "themeAccent": { "type": "string", "nullable": true },
          "soundEnabled": { "type": "boolean" }
        }
      },
      "AuthResponse": {
        "type": "object",
        "properties": {
          "user": { "$ref": "#/components/schemas/User" },
          "token": { "type": "string", "description": "Short-lived access token (15m)" },
          "accessToken": { "type": "string" },
          "refreshToken": { "type": "string", "description": "Long-lived refresh token (30d)" }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": { "type": "string" },
          "code": { "type": "string" },
          "timestamp": { "type": "number" }
        }
      }
    }
  },
  "paths": {
    "/health": {
      "get": {
        "summary": "Health check",
        "description": "Check service and database status",
        "responses": {
          "200": {
            "description": "Service is healthy",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": { "type": "string", "example": "ok" },
                    "environment": { "type": "string", "example": "development" },
                    "timestamp": { "type": "number" }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/auth/register": {
      "post": {
        "summary": "User registration",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["email", "password", "firstName"],
                "properties": {
                  "email": { "type": "string", "format": "email" },
                  "password": { "type": "string", "minLength": 6 },
                  "firstName": { "type": "string" },
                  "lastName": { "type": "string" },
                  "username": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Account created successfully",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/AuthResponse" }
              }
            }
          },
          "400": { "description": "Validation error or email already taken" },
          "429": { "description": "Too many registration attempts" }
        }
      }
    },
    "/auth/login": {
      "post": {
        "summary": "User login",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["email", "password"],
                "properties": {
                  "email": { "type": "string", "format": "email" },
                  "password": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Logged in successfully",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/AuthResponse" }
              }
            }
          },
          "400": { "description": "Invalid email or password" },
          "429": { "description": "Rate limit exceeded (15 attempts/15min)" }
        }
      }
    },
    "/auth/refresh": {
      "post": {
        "summary": "Exchange refresh token for a new access token",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["refreshToken"],
                "properties": {
                  "refreshToken": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Token rotated successfully",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/AuthResponse" }
              }
            }
          },
          "401": { "description": "Invalid or revoked refresh token" }
        }
      }
    },
    "/auth/logout": {
      "post": {
        "summary": "Log out user and revoke refresh token",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "refreshToken": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Logged out successfully" }
        }
      }
    },
    "/auth/me": {
      "get": {
        "summary": "Get current user profile",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "user": { "$ref": "#/components/schemas/User" }
                  }
                }
              }
            }
          },
          "401": { "description": "Unauthorized" }
        }
      }
    },
    "/sync/push": {
      "post": {
        "summary": "Push offline mutations to cloud database",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": { "description": "Mutations applied via Last-Write-Wins" },
          "400": { "description": "Validation error in push payload" },
          "401": { "description": "Unauthorized" }
        }
      }
    },
    "/sync/pull": {
      "get": {
        "summary": "Pull latest updates from cloud database",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "since",
            "in": "query",
            "required": false,
            "schema": { "type": "integer" },
            "description": "Timestamp in milliseconds of last client sync"
          }
        ],
        "responses": {
          "200": { "description": "Cloud updates since timestamp" },
          "401": { "description": "Unauthorized" }
        }
      }
    }
  }
}